[network] Rate-limit inbound messages
Currently, a peer can send any number of messages in a given time
window. To allow for fair-sharing of resources as well as DoS
prevention, we add a rate-limiter in front of message reader, which
applies back-pressure when the number of messages exceeds the window
limit.

This commit adds the rate-limit config as hard-coded constants, but we
may wish to expose these as configurable knobs in the future.