[move-prover] Added Boogie model for serialize to prelude.bpl
Move datastructures are converted to vector<u8> for sha2/3, crypto and other purposes by a native library called "serialize".  There is no "deserialize" in Move and no plans to create one.

The new Boogie code for serialize maps a Move value to an arbitrary non-empty vector<8>.  There is a an additional axiom that constrains serialize to be an injection, because (I believe) the correctness of Move modules will sometimes depend on this property. However, the model is a conservative approximation of the actual function because it does not guarantee that the bytes in the serialized value correspond to what the native serialize function does, or even that the length of the result is the same as the actual serialize.

We should be able to verify properties that require serialize to be an injection, but won't be able to verify code that accesses and manipulates bytes from a serialized value or, for example, code that depends on the length of serialized values.

There are test cases serialize-model.move and serialize-model-invalid.move, also. Some new .exp files were created for stdlib files that use lcs.move.