[move-lang] allow specifying directories for input files and dependencies
When compiling a Move file that is not part of the standard library, there
are typically a number of dependencies on the stdlib modules. It is a
hassle to specify these individually. This change lets you specify the
directory containing the stdlib modules as a dependency, and the compiler
will expand that to include all the ".move" files inside that directory.
We can do the same for the input files to make it easier to build the
stdlib (or any other collection of Move files).

This also changes the move-build and move-check programs to customize
the error printing. The default output for io::Error results is not
very pretty.