[safety-rules] Add ability to store waypoint
The next stream of work is to handle initializing safety rules from a
known point via a waypoint. So we add this first. Next is to build a
genesis-waypoint computer to retrieve this as part of config-builder.
Afterward, add the waypoint to the config. At which point, consensus and
safety-rules can communicate about the current waypoint and have
consensus pass an appropriate LI containing a validator set to
safety-rules along with replaying an appropriate validator change proof.