[config] dynamically derive public key from private key
We currently obtain both the private key and public key (of different types)
from config files.
We then keep them both in a struct without verifying that the public key
is indeed the correct one associated to the private counterpart.

This PR changes this behavior to only serialize and deserialize the
private key part. The public key part is dynamically derived during
deserialization of the private key.

This way, no mistakes can be made when given a configuration that has a
wrong public key.