[language] save a copy of the stdlib source files when staging
To compile a script or module that is not built with the stdlib,
the Move compiler needs to import the stdlib source files for any
dependencies. Those dependencies should match the Move binaries that
are stored on-chain, so we need to save a copy of those sources when
staging the stdlib binaries.