Use the Move source language for client_proxy and libratest/smoke_test
Some changes here may not be obvious:

- The compile_program function used to perform an automatic substitution
of "{{sender}}" in the code, but I moved that functionality to the caller
(e.g., smoke_test.rs). The module.move file needs an "address" directive
and that file also needs to be provided as a dependency when compiling
script.move. The temporary file with the substituted address was not
previously accessible, and I didn't want to blindly rewrite all the
dependency files. Doing the substitution in the caller makes it possible
to reuse the file.

- The source compiler does not support a feature like the IR compiler
has to list dependencies. Instead this relies on a new feature in the
source compiler (PR #3373) to include the entire Move stdlib as a
dependency.

- Removing the code to grab the dependency bytecode from on-chain made
the get_account_blob function in libra_client.rs unused, so I removed it
as well.