[move-prover] Generic spec vars and more features.
This implements generic specification variables which allow to track global state of generic resources. It also allows specification functions to reference spec vars, an essential feature missing before. For an e2e use case which mimics the marketcap specification problem in presence of generics, see the new test `tests/sources/marketcap_generic.move`.

In order to make this feature work, we attach the new type parameter declarations to a `SpecVarDecl` in the environment, and use this information in `spec_translator.rs` to generate accesses/updates to Boogie arrays instead of single values where needed. We also had to add the `TypeValue` parameters to various helper functions (pack, unpack, and update invariants). Finally, we detect spec vars used in spec functions and automatically add them as Boogie parameters so they can be resolved.