[types] Add NetworkAddress as multiaddr replacement
Libra `NetworkAddress` is a compact, efficient, self-describing and
future-proof network address represented as a stack of protocols.
Essentially libp2p's [multiaddr](https://multiformats.io/multiaddr/)
but using `lcs` to describe the binary format.

One additional motivation is to eliminate our dependency on
`parity-multiaddr`, which is apparently a weighty dependency and would
have to be included in our very-locked-down safety rules module.