[network] check timeout before doing anything
Right now the network timeout is checked after polling the inner future,
this leads to unecessary work.
At worse it can lead to sending data on the wire for nothing,
or parsing a response and computing cryptographic operations for nothing.

This PR reverses the logic to first check the timeout and then poll.