[genesis] Use scripts to initialize ValidatorConfigs
Historically, we would generate both ValidatorSet and DiscoverySet to
then produce the correct interface into executing functions within the
VM. This results in a lot of extra data being generated and an
interesting dependency on NodeConfigs. This is a first step toward
leveraging the proposed concept for genesis, which would be to leverage
signed transactions from the would-be validators on their accounts.

This is incremental in that the following need to happen:

* Move config::generator into its own crate, eliminate the generation of
discovery and validator sets, and generate the scripts there.
* Discuss with the move team on the best place for handling
TransactionArgument <-> Value
* Consider leveraging the LibraVM for verification rather than the MoveVM
* All the work that goes into deoupling the automated generation of
genesis from the code base and make it map to both.