[secure-storage] Add a default policy
This enables using create without actually specifying a policy, which
can be the case if a policy already exists for the location we are
storing to, which is quite common across multiple frameworks.

We actually should refactor the API to separate policy creation and data
creation, but that is a far bigger task that will need to be worked on
later.