[management] Add a layout config
Layout defines the set of roles to identities within genesis. In
practice, these identities will map to distinct namespaces where the
expected data should be stored in the deterministic location as defined
within this tool.

This only shows the process for adding the layout upstream future
commits will show how the layout will be used.