[move-prover] Replace "IsEqual" with "==" in Boogie prelude.bpl
In the prelude that is included in generated boogie code, we were
using a recursive function, IsEqual, to check equality.  Since a real
recursive function would have caused trouble in Boogie, it was
stratified to a fixed depth ("IsEqual1", "IsEqual2", etc.).

There were several problems with this beyond ugliness. SMT solvers
understand equality extremely well, including the substitution
property, which is built-in and handled efficiently. "IsEquals" was
just another function to them without special properties. This change
will, I hope, enable a more efficient model for hashing and
serialization, where I am now using an uncomfortable number of
quantified formulas (I will try that later).  Finally, it unexpectedly
reduced the size of generated SMT files by a factor of about 50-100
(although, not with dramatic speedups).

In order to make it work, I had to force Boogie to use the builtin SMT
theory of arrays, which is extensional (arrays are equal if all
elements are equal). Boogie's default theory is not extensional.
Finite length vectors and records are represented as SMT arrays (which
have infinite index sets) and a length. IsEqual only compared these
for indices 0 .. len-1, so that invalid entries weren't compared.  I
had to take a little more care that all vectors and structs had
(Error) (also known as DefaultValue, currently) at all invalid
indices.  In particular, I had to replace generated "assume
is#vector(v)" in Module_Fun_is_well_formed with
$Vector_is_well_formed, which not only makes sure v is a vector, but
that its invalid elements are all (Error). That has a quantified
formula in it, but it's worked reliably so far.

I've changed the command-line option --use-array-theory to
--no-array-theory.  If you use --no-array-theory, you should specify
the old prelude, -p src/no_array_theory_prelude, on the command line.
Once we are confident that the array theory works reliably (or as
reliably as the old version), we can get rid of this option.

I cleaned up both preludes, updating outdated comments, deleting unused
functions, and making spacing more consistent.