[language] Bring cost table code in the Move VM
This move cost table related code away from the VM and into move-core/types
and move-vm/types.
The goal is to bring everything about native functions into the
MoveVM where we can rationalize that first and then maybe relax and move
it to the Libra layer.
That cleaned up dependencies a bit too