[language, move] Introduce public key validation
Introduce a native function for public key validataion, and use it in key rotation.

This solves the problem that any 32-byte value would be accepted as a public key in the stdlib contracts, despite the length check being insufficient to protect against [invalid point attacks, small subgroup attacks](https://link.springer.com/content/pdf/10.1007/3-540-44598-6_8.pdf), and typos. A user would typically rotate their key to an invalid public key, wittingly or no, and then fail to produce signatures against that new key, losing the balance of their account in the process.

This introduces a native function for checking the validity of this public key, and uses it upon rotation. It would probably be useful to call this public key validation function in a few more places - in fact, anywhere externally-controlled key material is registered on-chain. I have not looked beyond the key rotation contracts for now.

A further step would be proof of ownership (TBD).