[easy] Change unwraps-in-a-map to collect-to-result
The transformation very strictly maintains the semantics of the original, i.e. panic on the first failure.
It demonstrates the proper way to collect faillible iterations into a Result, a [useful pattern](https://doc.rust-lang.org/stable/rust-by-example/error/iter_result.html#fail-the-entire-operation-with-collect) that should be more well-known.

This also has the advantage of enabling the use of the `?` operator in the functions that allow it (depending on their return type), and popping out the `.unwrap()` in those that do not, making the panic pattern easier to understand.