[move-prover] Quantification of addresses in specs
Prior to this extension, we had quantification over finite ranges and vectors.
So, you could say all(1..10, |i| i < 20) (universal quantification) or
any(1..10, |i| < 3) (existential quantification).

This extension allows formulas to be quantified over all addresses. For
example, the property:
    all(addresses(), |a| exists<R1>(a) ==> !exists<R2>(a))
says that no address with an R1 resource may have an R2 resource,
and
    exists(addresses(),  exists<R1>(a))
says that there is some address with an R1 resource.

Such properties are often useful in module invariants to assert that a
top-level resource exists at at most one address, that top-level
resources are mutually exclusive, that if there is resource R1, there
is always a resource R2 at the same address, etc.

IMPLEMENTATION:
I added a new type "addresses" representing the set of all addresses, and
a new native function addresses().  When the translator encounter an "any"
or "all" where the first argument is the "addresses" type, e.g.
    all(addresses(), |a| P(a)), it generates Boogie code:
    forall a : Value :: is#address(a) ==> <boogie for P(a)>
For any(addresses(), |a| P(a)), it generates Boogie code:
    exists a : Value :: is#address(a) && <boogie for P(a)>

There is a new test file
tests/sources/functional/address_quant.move