[move-prover] Align the command-line syntax with the move-lang compiler
Rename the "dep" option to "dependency" and remove the "search-path"
option. Allow either the source file or dependency option to specify a
directory that is recursively searched for ".move" files.

For the compiler, it is not too expensive to pull in unused dependencies,
but that would make the prover too slow. This keeps the prover code that
scans for module dependencies and uses that to prune out unused
dependencies.

Closes: #3970
Approved by: wrwg