[move] Introducing module sliding_nonce
Sliding nonce is a resource, that can be used in custom scripts to allow for transaction deduplication with limited out-of-order execution.

Sliding keeps fixed size state allowing up to 128 out-of-order transaction to exist at any given time.

For sliding nonce to work effectively, nonce generator for transaction should produce nonce in sequential manner.