[move-prover] More global specifications
This contains additional specifications in association.move, with better
doc formatting.

There is also a change in the code to (1) allow "requires module" to
be added to private functions (in spec-lang/src/ast.rs) , and (2) to
add assumptions to the _verify function wrapper for "requires" for
private as well as public functions (in src/spec_translator.rs).

Specifications in libra.move have changed, and there are new
specifications in register_currency.move and libra_configs.move.

Verification has been disabled in libra.move, libra_configs.move, and
registered_currencies.move because of various problems -- I want to
share these specs without more delay.