[consensus] remove special handling of NIL block for round interval
We use to not count NIL block as committed to avoid shrinking the round
interval too short in which we can't execute a big block and continue to
commit NIL blocks forever.

Although we don't interrupt the execution upon timeout, it's possbile
to vote even execution takes longer than the round interval.