[network] Remove signed notes from gossip
We only have gossip so that we can instantiate networks on the fly --
for test purposes. We had planned to remove gossip, but there will be a
slow movement from gossip to onchain discovery because test networks --
work so well... So we are left with a dilema. Do we maintain the
complexity of the existing gossip network, which impacts our network
configuration substantially -- just for a testing api or do we
substantially simplify the testing framework to not impact the
production environment?

This PR says tests will be tests and compromises towards simplifying our
production deployment. With that in mind, there are no more verified
notes. gossip is "insecure".

What does it mean for gossip to be insecure? Well, some really evil
sneaky peer (on a test network) could overwhelm us with peers that we
don't know about or eliminate our ability to leverage gossip from other
peers. That's not such a big issue:
1) Testnetworks are private to a single organization so passing around
these gossip messages is ... not possible.
2) If we are concerned about some sort of attack on overwhelming us, we
could substnatially improve this -- and we can -- it isn't production
code, so let's not over think this.

By doing this, we can eliminate signing keys from all configuration,
which will make our genesis construction so much simpler too...

Closes: #4146
Approved by: gregnazario