[config] Enable a new enum for handling keys for networking
We need to either load the key from config or from storage. This exposes
the basic logic for this configuration and begins by retaining the
existing behavior of from config. Its important to note that this
configuration really doesn't address the long term desire for enabling
automated rotation, but that's a long term goal ... certainly not an
immediate goal.