[move core modules] VASP account recovery module
This PR introduces a new resource that holds the KeyRotationCapability's for several accounts belonging to the same VASP. A VASP account that delegates its KeyRotationCapability to a RecoveryAddress resource retains the ability to rotate its own authentication key, but also allows the account that stores the RecoveryAddress resource to rotate its authentication key.

This is useful as an account recovery mechanism: VASP accounts can all delegate their rotation capabilities to a single `RecoveryAddress` resource stored under address A. The authentication key for A can be "buried in the mountain" and dug up only if the need to recover one of accounts in rotation_caps arises.