[network] error on receiving a noise handshake payload in FN network
In a full node network, or as we call it currently a ServerOnly handshake mode,
we do not send a payload during the handshake.
(In the validator network we do, for anti-replay mitigation.)
Thus, we should check on the responder/server side that the payload was empty,
and return an error otherwise.

Closes: #4325
Approved by: phlip9